#6 [refactor] : navigation 정리
