+++ app/src/main/java/org/sopt/and/feature/login/LoginViewModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.and.feature.login
+
+import androidx.lifecycle.ViewModel
+
+class LoginViewModel : ViewModel() {
+}
