+++ app/src/main/java/org/sopt/and/feature/search/SearchViewModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.and.feature.search
+
+import androidx.lifecycle.ViewModel
+
+class SearchViewModel : ViewModel() {
+}
